+++ src/gaming/ex12/Bird.java
@@ -5,7 +5,7 @@ import javafx.scene.canvas.GraphicsContext;
 
 public class Bird extends Player implements Enemy {
 	public static final int INITIAL_Y = 100;
-	public static final int INITIAL_X = 300;
+	public static final double INITIAL_X = 300;
 	public static final int WANDER_WIDTH = 300;
 	private RunAnimation flyingAnimation = new RunAnimation(ImageResource.BIRD);
 	private RunAnimation fallingAnimation = new RunAnimation(ImageResource.BIRD, 20, 2);
